requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-vision==3.4.5
google-generativeai==0.3.2
Werkzeug==2.2.3
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0
Flask-Session==0.5.0
